refactor improvements
