+++ frontend_cms/src/App.tsx
@@ -1,24 +1,35 @@
-import "./App.css";
+import "App.css";
 import "bootstrap/dist/css/bootstrap.min.css";
 import { QueryClient, QueryClientProvider } from "react-query";
-import Issues from "./Issues";
+import Issues from "Components/Issues/Issues";
 import { Container } from "react-bootstrap";
-import CongressionalBills from "./CongressionalBills";
+import CongressionalBills from "Components/CongressionalBills/CongressionalBills";
+import { useState } from "react";
+import { ModalContext } from "hooks/useModal";
 
 const queryClient = new QueryClient();
 
 function App() {
+    const [modal, setModal] = useState<JSX.Element | undefined>(undefined);
     return (
         <div className="App">
-            <Container>
-                <QueryClientProvider client={queryClient}>
-                    <h1 className="mb-4 mt-4 text-white">
-                        Climate Changemakers CMS
-                    </h1>
-                    <Issues />
-                    <CongressionalBills />
-                </QueryClientProvider>
-            </Container>
+            <QueryClientProvider client={queryClient}>
+                <ModalContext.Provider
+                    value={{
+                        open: (modal: JSX.Element) => setModal(modal),
+                        close: () => setModal(undefined),
+                    }}
+                >
+                    <Container>
+                        <h1 className="mb-4 mt-4 text-white">
+                            Climate Changemakers CMS
+                        </h1>
+                        <Issues />
+                        <CongressionalBills />
+                    </Container>
+                    {modal}
+                </ModalContext.Provider>
+            </QueryClientProvider>
         </div>
     );
 }
