+++ frontend_cms/src/hooks/useCreateBillMutation.tsx
@@ -1,0 +1,9 @@
+import { useMutation } from "react-query";
+import { fetcher } from "../helper";
+import { Bill } from "../Types/Bill";
+
+export default function useCreateBillMutation() {
+    return useMutation<undefined, Error, Bill>((formData) =>
+        fetcher("bills", formData)
+    );
+}
