+++ frontend_cms/src/hooks/useCreateIssueMutation.tsx
@@ -1,0 +1,10 @@
+import { useMutation } from "react-query";
+import { fetcher } from "../helper";
+import { ExistingIssue, IssueInfo } from "../Types/Issue";
+
+export default function useCreateIssueMutation() {
+    return useMutation<ExistingIssue | undefined, Error, IssueInfo, unknown>(
+        (formData: IssueInfo) =>
+            fetcher<ExistingIssue | undefined, IssueInfo>("issues", formData)
+    );
+}
