+++ frontend_cms/src/hooks/useUpdateBillsForIssueMutation.tsx
@@ -1,0 +1,9 @@
+import { useMutation } from "react-query";
+import { fetcher } from "../helper";
+
+export default function useUpdateBillsForIssueMutation() {
+    return useMutation<unknown, Error, { issueId: number; billIds: number[] }>(
+        (formData) =>
+            fetcher(`issues/${formData.issueId}/bills`, formData, "PUT")
+    );
+}
