+++ frontend_cms/src/index.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import "./index.css";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
+import "index.css";
+import App from "App";
+import reportWebVitals from "reportWebVitals";
 
 const root = ReactDOM.createRoot(
     document.getElementById("root") as HTMLElement
